Fix bugs and clean up
There was a problem with the way the AI selected the next move,
such that it would always start with the highest possible move
on the turn (i.e. 9 or 8). Fixed w. indentation.

There was a problem where the AI would update the states_value for
all states it had seen, rather than all states that game. This lead
to long (and increasing) training times and wrong learning outcomes.

Cleanup: Removed comments, code for bug fixing and stale code.